requirements.txt
Django==3.1.3
djangorestframework==3.12.2
pyscopg2
django-cors-headers